refactor:  change search film page
