+++ client-app/src/containers/FilmSearchPage.js
@@ -1,6 +1,6 @@
 import React, {Component} from 'react';
 import '../components/Films/film.scss';
-import FilmInfoContainer from '../components/Films/FilmInfoContainer';
+import FilmList from '../components/Films/FilmList';
 import SearchBar from '../components/Films/SearchBar';
 import Header from '../components/Authentication/Header';
 
@@ -11,12 +11,9 @@ class FilmSearchPage extends Component {
     return (
       <section className="page-content">
         <div className="container">
-          {/*<div className="row">
-            <h2 className="authentication__header">Film Search</h2>
-    </div>*/}
           <Header header="Film Search"/>
           <SearchBar />
-          <FilmInfoContainer />
+          <FilmList />
         </div>
       </section>
     )
